refactored auth dtos
