+++ shared/models/auth/SignInReqDto.ts
@@ -1,0 +1,4 @@
+export interface SignInReqDto {
+    username: string,
+    password: string,
+}
