+++ shared/models/auth/SignInResDto.ts
@@ -1,0 +1,3 @@
+export interface SignInResDto {
+    access_token: string
+}
