+++ shared/models/auth/SignUpDto.ts
@@ -1,0 +1,5 @@
+export interface SignUpDto {
+    username: string,
+    password: string,
+    email: string
+}
